backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
boto3==1.34.129
python-dotenv==1.0.1
Pillow==12.1.0
pytest==9.0.2
pytest-flask==1.3.0
requests==2.31.0